Deleted extra folder n added assertions for login
