+++ CustomReporterConfig.ts
