+++ pageFactory/pageRepository/LoginPage.ts
@@ -2,7 +2,7 @@ import {LoginPageObjects} from "@objects/LoginPageObjects"
 import type { Page } from '@playwright/test';
 import {testConfig} from '../../testConfig';
 
-export default class LoginPage  {
+export class LoginPage  {
 
 public page:Page;
 public loginpageobjects:LoginPageObjects
@@ -11,18 +11,17 @@ constructor(page:Page,loginpageobjects:LoginPageObjects) {
      this.page= page;
      this.loginpageobjects=loginpageobjects;
 }
-async navigationtourl(){
+async loginToHrmsWithValidCredentials(username,password) {
     await this.page.goto(testConfig.url)
-}
-async loginToApplication(username,password) {
     await this.loginpageobjects.login_username().fill(username)
     await this.loginpageobjects.login_password().fill(password)
     await this.loginpageobjects.login_btn().click() 
-}
-async AddSkill_cancel(){
     await this.loginpageobjects.Addskill_cancel().click()
 }
-async cleartext(){
+async loginToHrmsWithInValidCredentials(username,password) {
+    await this.loginpageobjects.login_username().fill(username)
+    await this.loginpageobjects.login_password().fill(password)
+    await this.loginpageobjects.login_btn().click() 
     await this.loginpageobjects.login_username().fill("")
     await this.loginpageobjects.login_password().fill("")
 }
