+++ pageFactory/pageRepository/LogoutPage.ts
@@ -1,16 +1,12 @@
 import {LogoutPageObjects} from "@objects/LogoutPageObjects"
 import type { Page } from '@playwright/test';
 
+export class LogoutPage  {
 
-export default class LoginPage  {
-
-constructor(page:Page,logoutpageobjects:LogoutPageObjects) {
-     //this.page=page;
-     this.logoutpageobjects=logoutpageobjects;
-}
-
-//private page:Page;
-private logoutpageobjects:LogoutPageObjects
+ constructor(page:Page,logoutpageobjects:LogoutPageObjects) {
+    this.logoutpageobjects=logoutpageobjects;
+  }
+   public logoutpageobjects:LogoutPageObjects
 
     async Logout() {
         await this.logoutpageobjects.Profile().click()
