--- playwright-typescript-playwright-test-main/playwright-typescript-playwright-test-main/tests/functional/RaiseTicket.test.ts
@@ -1,20 +1,0 @@
-//* /import test from '@lib/BaseTest';
-
-
-//test(`Go to RaiseTicket`, async ({ loginPage, RaiseTicketPage,LogoutPage }) => {
-
-    //await loginPage.navigateToURL();
-    //await loginPage.loginToDotnetEmployee();   
-    //await RaiseTicketPage.Raise_A_Ticket();
-    //await LogoutPage.Logout();
-    //await loginPage.loginToAdminHod();
-    //await RaiseTicketPage.Resolve_Ticket();
-
-//})
-     
-   // */
-    
-    
-   
-
-
